Equality of Covariance Matrices

Bartlett's test

Bartlett's test for equality of two variance-covariance matrices is provided. This is equivalent to Box's $M$-test for two groups.

API

MultivariateTests.BartlettsTest — Type.

BartlettsTest(X::AbstractMatrix, Y::AbstractMatrix)

Perform Bartlett's test of the hypothesis that the covariance matrices of X and Y are equal.

Note

Bartlett's test is sensitive to departures from multivariate normality.

source