Equality of Mean Vectors

Hotelling's $T^2$ test

The HypothesisTests framework is extended here to include Hotelling's $T^2$ test for one and two samples.

API

MultivariateTests.OneSampleHotellingT2 — Type.

OneSampleHotellingT2(X::AbstractMatrix, μ₀=<zero vector>)

Perform a one sample Hotelling's $T^2$ test of the hypothesis that the vector of column means of X is equal to μ₀.

source

OneSampleHotellingT2(X::AbstractMatrix, Y::AbstractMatrix, μ₀=<zero vector>)

Perform a paired Hotelling's $T^2$ test of the hypothesis that the vector of mean column differences between X and Y is equal to μ₀.

source

MultivariateTests.EqualCovHotellingT2 — Type.

EqualCovHotellingT2(X::AbstractMatrix, Y::AbstractMatrix)

Perform a two sample Hotelling's $T^2$ test of the hypothesis that the difference in the mean vectors of X and Y is zero, assuming that X and Y have equal covariance matrices.

source

MultivariateTests.UnequalCovHotellingT2 — Type.

UnequalCovHotellingT2(X::AbstractMatrix, Y::AbstractMatrix)

Perform a two sample Hotelling's $T^2$ test of the hypothesis that the difference in the mean vectors of X and Y is zero, without assuming that X and Y have equal covariance matrices.

source